+++ src/components/services/CryptoTrackerPanel.tsx
@@ -1,5 +1,9 @@
 "use client";
 
+import { IframePanel } from "./IframePanel";
+
+const SRC = process.env.NEXT_PUBLIC_CRYPTO_TRACKER_URL ?? "https://192-18-128-170.nip.io";
+
 export function CryptoTrackerPanel() {
-  return null;
+  return <IframePanel name="Crypto Tracker" src={SRC} />;
 }